Delete compiled assets directory before building
Deleting the src/compiled-assets directory allows us to account for the case when src/compiled-assets contains a file no longer needed (for instance, a file that Webpack used to build but no longer builds). If we do not delete it, Eleventy will copy it to the dist/assets directory (since it copies the entire compiled-assets directory to dist/assets), and our website will end up having an unnecessary file.